+++ apps/payments/src/domain/schema.ts
@@ -1,5 +1,14 @@
 import { integer, jsonb, pgSchema, text, timestamp, unique, uuid } from "drizzle-orm/pg-core";
 
+import type { OrderRecord } from "@tix/contracts/orders";
+import type { PaymentIntentStatus } from "@tix/contracts/payments";
+
+// The read-model mirrors the Order aggregate's status, so it speaks the same closed set
+// the orders contract defines rather than a bare `string`. Today payments only projects
+// `created`/`cancelled`, but deriving from the contract keeps the column honest if more
+// order.* events get consumed later.
+export type OrderReadModelStatus = OrderRecord["status"];
+
 export const paymentsSchema = pgSchema("payments");
 
 export const payments = paymentsSchema.table("payments", {
@@ -11,7 +20,7 @@ export const payments = paymentsSchema.table("payments", {
   stripeId: text("stripe_id").notNull().unique(),
   amount: integer("amount").notNull(),
   currency: text("currency").notNull(),
-  status: text("status").notNull(),
+  status: text("status").$type<PaymentIntentStatus>().notNull(),
   version: integer("version").notNull().default(1),
   createdAt: timestamp("created_at", { withTimezone: true }).notNull().defaultNow(),
   updatedAt: timestamp("updated_at", { withTimezone: true })
@@ -25,7 +34,7 @@ export const orderReadModel = paymentsSchema.table("order_read_model", {
   version: integer("version").notNull(),
   userId: text("user_id").notNull(),
   priceCents: integer("price_cents").notNull(),
-  status: text("status").notNull(),
+  status: text("status").$type<OrderReadModelStatus>().notNull(),
   createdAt: timestamp("created_at", { withTimezone: true }).notNull().defaultNow(),
   updatedAt: timestamp("updated_at", { withTimezone: true })
     .notNull()
